corre2

+++ shared/src/commonMain/kotlin/com/sdevprem/runtrack/shared/domain/tracking/TrackingManager.kt
@@ -84,7 +84,7 @@ private fun addPathPoints(info: LocationTrackingInfo) {
                     distance += getSafeDistance(
                         pathPoints[pathPoints.size - 1],
                         pathPoints[pathPoints.size - 2]
-                    )
+                    ).toInt()
                 }
                 distance
             },
